fix: set Id as UUID and Long?

+++ src/main/kotlin/ru/aps/performance/controllers/ChatRoomController.kt
@@ -7,6 +7,8 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.GetMapping
 import ru.aps.performance.services.ChatRoomService
 import ru.aps.performance.dto.ChatRoomRequest
+import ru.aps.performance.exceptions.NotEnoughParticipantsException
+import ru.aps.performance.exceptions.NoSuchChatRoomException
 import java.util.UUID
 
 @RestController
@@ -15,6 +17,9 @@ class ChatRoomController(
 ) {
     @PostMapping("/chatRoom")
     fun addChatRoom(@RequestBody chatRoomRequest: ChatRoomRequest) {
+        if (chatRoomRequest.secondUserId == null) {
+            throw NotEnoughParticipantsException("Not enough participants to create a chat between")
+        }
         val firstUserId = UUID.fromString(chatRoomRequest.firstUserId)
         val secondUserId = UUID.fromString(chatRoomRequest.secondUserId)
         chatRoomService.addChatRoom(firstUserId, secondUserId, chatRoomRequest.name)
@@ -22,6 +27,9 @@ class ChatRoomController(
 
     @DeleteMapping("/chatRoom")
     fun deleteChatRoom(@RequestBody chatRoomRequest: ChatRoomRequest) {
+        if (chatRoomRequest.chatRoomId == null) {
+            throw NoSuchChatRoomException("Not valid chatRoomId or firstUserId")
+        }
         val chatRoomId = UUID.fromString(chatRoomRequest.chatRoomId)
         val firstUserId = UUID.fromString(chatRoomRequest.firstUserId)
         chatRoomService.deleteChatRoom(chatRoomId, firstUserId)
@@ -29,6 +37,9 @@ class ChatRoomController(
 
     @GetMapping("/chatRoom")
     fun getChatRoom(@RequestBody chatRoomRequest: ChatRoomRequest) {
+        if (chatRoomRequest.chatRoomId == null) {
+            throw NoSuchChatRoomException("Not valid chatRoomId or firstUserId")
+        }
         val chatRoomId = UUID.fromString(chatRoomRequest.chatRoomId)
         val firstUserId = UUID.fromString(chatRoomRequest.firstUserId)
         chatRoomService.getChatRoom(chatRoomId, firstUserId)
